Further improvements:
1) The parsing logic moved into ApplicationParser
2) String Array IsNullOrEmpty method has been removed an instead such checking was built into the method.
3) ToolBoxOptionsConfiguration files has been moved from QC.Toolbox to QC.Configuration project
